ClearChannels change

+++ src/main/java/com/beyondbell/bugisoft/Commands/Lobby/ClearChannels.java
@@ -6,6 +6,10 @@ public class ClearChannels {
 
 	//event pushed from 'VoiceChannelLeaveEvent', conditions are met in the class
 	public ClearChannels(GuildVoiceLeaveEvent event) {
-		event.getChannelLeft().delete();
+		if(!event.getChannelLeft().getName().toLowerCase().equals("spotify")
+				|| !event.getChannelLeft().getName().toLowerCase().equals("no game lobby")
+				&& event.getChannelLeft().getMembers().size() == 0) {
+			event.getChannelLeft().delete();
+		}
 	}
 }
